requirements.txt
# pip install -r requirements.txt

# base ----------------------------------------
matplotlib>=3.4.2
numpy>=1.19.2
opencv-python>=4.4.0.46
Pillow>=8.3.1
scipy>=1.6.3
torch==1.7.1+cu110
torchvision==0.8.2+cu110
tqdm>=4.48.2